+++ core/src/message/handle.rs
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::identifiers::{MessageId, ProgramId};
+use crate::ids::{MessageId, ProgramId};
 use crate::message::{
     Dispatch, DispatchKind, GasLimit, Message, Payload, StoredDispatch, StoredMessage, Value,
 };
