+++ pallets/usage/src/mock.rs
@@ -22,7 +22,7 @@ use common::Origin as _;
 use frame_support::traits::{ConstU64, FindAuthor, OffchainWorker, OnInitialize};
 use frame_support::{construct_runtime, parameter_types};
 use frame_system as system;
-use gear_core::{identifiers::CodeId, program::Program};
+use gear_core::{ids::CodeId, program::Program};
 use parking_lot::RwLock;
 use primitive_types::H256;
 use sp_core::offchain::{
